add: moved and renamed card component for mypage
